# perf(ops): support kBool return type and kV8Value argument type in fast API ops