--- a/cli/bench/deno_common.js
+++ b/cli/bench/deno_common.js
@@ -47,6 +47,14 @@ Deno.bench("open_file_sync", () => {
   file.close();
 });
 
+{
+  const { rid } = Deno.openSync("./cli/bench/testdata/128k.bin");
+  const { tryClose } = Deno.core;
+  Deno.bench("try_close_rid", () => {
+    tryClose(rid);
+  });
+}
+
 // A common "language feature", that should be fast
 // also a decent representation of a non-trivial JSON-op
 {
@@ -94,3 +102,18 @@ Deno.bench(
 );
 
 Deno.bench("request_new", { n: 5e5 }, () => new Request("https://deno.land"));
+
+Deno.bench("response_new_headers", () => new Response(null, { 
+  headers: {
+    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
+    "Accept-Encoding": "gzip, deflate, br",
+    "Accept-Language": "en-US,en;q=0.9",
+    "Cache-Control": "max-age=0",
+    "Connection": "keep-alive",
+  }
+}));
+
+const { isProxy } = Deno.core;
+Deno.bench("is_proxy", () => {
+  isProxy({});
+})