This makes `core.tryClose` & `core.isProxy` 2-3x faster. 

There are also visible improvements in webidl Request/Response headers converter and Deno.openSync benchmark.

```
main:
  try_close_rid          54.47 ns/iter   (52.79 ns ... 60.62 ns)  54.55 ns  59.36 ns  60.22 ns
  is_proxy               29.26 ns/iter   (27.21 ns ... 33.88 ns)  30.83 ns  31.19 ns  31.32 ns
  response_new_headers  911.85 ns/iter (906.53 ns ... 944.67 ns) 912.07 ns 944.67 ns 944.67 ns

this patch:
  try_close_rid          37.55 ns/iter   (37.3 ns ... 67.65 ns)   37.38 ns  39.46 ns  39.91 ns
  is_proxy               10.81 ns/iter    (9.57 ns ... 15.75 ns)  12.78 ns  13.22 ns  13.28 ns
  response_new_headers  881.89 ns/iter (874.26 ns ... 914.93 ns) 884.52 ns 914.93 ns 914.93 ns
```