@aapoalas 2022-08-21
Should the "slow call" `rv.set_bool(result)` optimised path also be enabled?