@aapoalas 2022-08-21
Tests fail from `null` not throwing as a parameter... There was a report on Discord about FFI APIs apparently managing to use an `undefined` value as a pointer without `serde_v8` throwing an error about that. Is it possible that both there and here the `null` / `undefined` argument is getting coerced to a 0?

For the `resourcesCloseBadArgs` case that would then close the `stdin` successfully, thus not throwing on the JS side and leading to the test failing.