@littledivy 2022-08-30
Superseded by https://github.com/denoland/deno/pull/15682